Read Raycast 2.x exports and retire the Raycast X beta reader
The beta that sat between Raycast 1.x and 2.x is over, so its gzip-envelope
reader is deleted rather than carried, and the RAYCFG3 container takes the v2
slot. Each format owns its own decrypt and its own mapper again, matching the
split v1 already used.

Also ignore *.rayconfig: a real export carries the author's clipboard history.
