+++ Scripts/run-tests.sh
@@ -149,8 +149,7 @@ run notes-editor-test      Tinycast/Platform/Signposts.swift \
                            Tinycast/Features/Notes/UI/NoteEditorView.swift
 run raycast-test           Tinycast/Features/Backup/Model/RaycastFormat.swift \
                            Tinycast/Features/Backup/Model/RaycastV1Decoder.swift \
-                           Tinycast/Platform/Hex.swift \
-                           Tinycast/Features/Backup/Service/RaycastV3Decoder.swift \
+                           Tinycast/Features/Backup/Service/RaycastV2Decoder.swift \
                            Tinycast/Features/Backup/Service/Scrypt.swift \
                            Tinycast/Platform/Compression/Zlib.swift \
                            Tinycast/Features/Clipboard/Model/ClipboardStore.swift \
