+++ Tests/snippets-test.swift
@@ -45,12 +45,12 @@ struct SnippetsTests {
 
     private static func testRaycastImport() {
         let imported = RaycastSnippetImport.parse([
-            ["name": "Email", "text": "person@example.com", "keyword": "  !email  "],
-            ["name": "Multiline 雪", "text": "First\nSecond"],
-            ["name": "Blank Keyword", "text": "Body", "keyword": "   "],
-            ["title": "Raycast 2", "text": "Current format"],
-            ["name": "   ", "text": "Skipped"],
-            ["name": "Missing Text"]
+            ["title": "Email", "text": "person@example.com", "keyword": "  !email  "],
+            ["title": "Multiline 雪", "text": "First\nSecond"],
+            ["title": "Blank Keyword", "text": "Body", "keyword": "   "],
+            ["title": "   ", "text": "Skipped"],
+            ["title": "Missing Text"],
+            ["name": "Retired Key", "text": "Skipped"]
         ])
 
         check(
@@ -58,9 +58,9 @@ struct SnippetsTests {
             RaycastSnippetImport.parse(["snippets": []]).isEmpty)
         check(
             "Raycast import keeps valid entries and source order",
-            imported.map(\.name) == ["Email", "Multiline 雪", "Blank Keyword", "Raycast 2"])
+            imported.map(\.name) == ["Email", "Multiline 雪", "Blank Keyword"])
         // The remaining assertions index into the result, so a wrong count has to fail rather than trap.
-        guard imported.count == 4 else { return }
+        guard imported.count == 3 else { return }
         check("Raycast import preserves text and Unicode", imported[1].text == "First\nSecond")
         check(
             "Raycast import trims keywords and normalizes blanks",
