--- Tinycast/Features/Backup/Service/RaycastImportV3.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-/// Keeps Raycast 2.x's container reader separate from Raycast X's envelope.
-enum RaycastImportV3 {
-    static func read(_ raw: Data, passphrase: String) throws -> RaycastImport.Result {
-        let payload = try RaycastV3Decoder.decrypt(raw, passphrase: passphrase)
-        return try RaycastPayloadMapper.result(from: payload, snippets: .current)
-    }
-}
