+++ docs/features/snippets.md
@@ -55,9 +55,8 @@ every change.
 ## Importing from Raycast
 
 The encrypted `.rayconfig` flow in **Settings → Backup** can import Raycast's built-in snippets as
-an independently selectable category. Tinycast reads `name` in older exports or `title` in Raycast 2,
-plus `text` and the optional `keyword`. These live in `builtin_package_snippets.snippets` in older
-exports and `snippets.snippets` in Raycast 2. Invalid entries are skipped; valid entries are added in
+an independently selectable category. Tinycast reads `title`, `text` and the optional `keyword` from
+the export's `snippets.snippets` collection. Invalid entries are skipped; valid entries are added in
 source order without overwriting the existing library. Duplicate names receive the same filename
 suffixes as snippets created in Tinycast, and duplicate keywords are preserved.
